OldBot/requirements.txt
Pillow==6.1.0
SpeechRecognition==3.8.1
asyncpg==0.18.3
dblpy==0.3.2
discord.py==1.2.3
granitepy==0.0.20a0
jishaku==1.15.0.134
matplotlib==3.1.1
objgraph==3.4.1
psutil==5.6.3
speedtest-cli==2.1.1
youtube-dl==2019.8.13